fix: python documentation refers to `projenrc.js` instead of `.projenrc.py` (#1695)
When the documentation was created it was not possible
to write the project setting in python file. Now, no
longer is necessary to use .projenrc.js as .projenrc.py
is available.

---
By submitting this pull request, I confirm that my contribution is made under the terms of the Apache 2.0 license.