feat(publisher): match first changelog entry (#1750)
The first ever changelog entry created by standard-version
doesn't link to a diff because none exists yet. For that
reason there are no markdown link brackets around the
version, which was causing the first manual release to
fail.

Closes #1749

---
By submitting this pull request, I confirm that my contribution is made under the terms of the Apache 2.0 license.