chore: pull out generic `Step` from `JobStep` (#1787)
This doesn't change the API, but moves generic `Step` properties out of `JobStep`. This is necessary because the [syntax](https://docs.github.com/en/actions/creating-actions/metadata-syntax-for-github-actions) for Action Metadata steps is ever so slightly different than a job step.

---
By submitting this pull request, I confirm that my contribution is made under the terms of the Apache 2.0 license.