fix: cannot create `PythonProject` with `.projenrc.js` (#1809)
Fixes #1807

To manage a PythonProject with a JavaScript projenrc file, the project will need to have these options configured:
```js
projenrcJs: true,
```

Or pass the option in the command line when creating the project, e.g. `npx projen new python --projenrc-js`.

---
By submitting this pull request, I confirm that my contribution is made under the terms of the Apache 2.0 license.