feat: `tryRemoveFile` escape hatch (#1819)
Adds an escape hatch method to all projects that allows removing files.

In the future, if projen uses the `constructs` programing model, this probably won't be needed since constructs have their own `node.tryRemoveChild` method - but for now this should be able to substitute for most use cases.

---
By submitting this pull request, I confirm that my contribution is made under the terms of the Apache 2.0 license.