chore: pin yaml in projen (#1856)
The new version of yaml (2.1.0) uses some types that are not supported in TypeScript 3.9, so projen fails to compile if we upgrade to it. Pin to the existing version for now.

---
By submitting this pull request, I confirm that my contribution is made under the terms of the Apache 2.0 license.