chore: fix flakey projen tests by decreasing jest worker count (#1816)
In #1809 I saw one of the builds fail due to a spurious test error that was unrelated to the changes made in the pull request:

```
FAIL test/release/bump.test.ts (755.713 s)
  ● first release

    thrown: "Exceeded timeout of 300000 ms for a test.
    Use jest.setTimeout(newTimeout) to increase the timeout value, if this is a long-running test."

      10 | jest.setTimeout(1000 * 60 * 5); // 5min
      11 |
    > 12 | test("first release", async () => {
         | ^
      13 |   const result = await testBump();
      14 |   expect(result.version).toStrictEqual("0.0.0");
      15 |   expect(result.changelog).toMatch(/.*## 0\.0\.0 \(\d{4}-\d{2}-\d{2}\).*/); // ## 0.0.0 (2021-01-01)

      at Object.<anonymous> (test/release/bump.test.ts:12:1)
```

If you run this test by itself on a development machine, it typically takes less than 10 seconds to run -- it should not take 5 minutes. After some debugging, I believe what may be the cause of this is that child processes created by some of these tests may be hanging due to CPU starvation / not getting enough CPU time.

I don't think it necessarily makes sense to make this setting the default in projen since many projects may be using jest without creating child processes so they may not run into this error.

---
By submitting this pull request, I confirm that my contribution is made under the terms of the Apache 2.0 license.