fix: java packaged exported with submodules was rending incorrect imp… (#1870)
When exporting a namespace in TS as follows: `export * as foo_bar from"./foo.ts"`, JSII will preserve the namespace when writing the .jsii file as it is not language specific. When jsii-packmak generates Java bindings, it removes the underscore so the package would be `my.org.foobar.Component`, however projen simply concats the namespace with the name which yields: `my.org.foo_bar.Component` which is not correct.

This PR aims to resolve this issues by looking up the packagename contained within the submodule type if found, otherwise falling back to the existing logic.

*Apologies for multiple commits - branch was outdated, hence the recent build failures*

---
By submitting this pull request, I confirm that my contribution is made under the terms of the Apache 2.0 license.

---
By submitting this pull request, I confirm that my contribution is made under the terms of the Apache 2.0 license.