fix!: update GitHub actions to recent versions (#1890)
Update several GitHub actions vended automatically by projen to their latest versions. Several of these include major version bumps so I've marked this as a breaking change for projen.

I wasn't able to test any of the changes since I'm not aware of a good way to do so. One way might be to have some dummy repos that run through these workflows live, but that seems tricky to maintain (and keep comprehensive)...

---
By submitting this pull request, I confirm that my contribution is made under the terms of the Apache 2.0 license.