chore: remove bump files from npm tarball (#1912)
Currently, the `package` task for jsii projects running in CI environments does the following:

```console
rsync -a . dist --exclude .git --exclude node_modules
```

It does this to prepare the `dist` directory for subsequent `package:${lang}` in the GitHub workflow. (More details [here](https://github.com/projen/projen/pull/1631))

The problem is that when this command is executed, the `dist` directory is not empty, it contains:

- `releasetag.txt`
- `version.txt`
- `changelog.md`

These are created in the `bump` task. This in turn creates a weird structure after the `rsync` is executed:

```console
dist/
├─ releasetag.txt // duplicated and misplaced
├─ version.txt // duplicated and misplaced
├─ changelog.md // duplicated and misplaced
├─ package.json
├─ lib/
├─ dist/
│  ├─ releasetag.txt
│  ├─ version.txt
│  ├─ changelog.md
```

Eventually the misplaced files end up in the published tarball. Its not horrible, but also undesirable. 

This PR fixes this quirk by `rsync`ing into an intermediary directory before populating the `dist` folder.

---
By submitting this pull request, I confirm that my contribution is made under the terms of the Apache 2.0 license.