fix(react): ReactTypeScript project not synthing sample code (#1918)
This is caused by: https://github.com/projen/projen/commit/41c95b73c06ef71bfdd2d146ad9f6263bdbb6db8.

The issue is the reactTypeDef is being synthed before the sampleDir and as such the sample code tries to synth in a non-empty dir which skips this.

This PR simplifies the process by just treating reactTypeDef as just a regular sample file and as such will be synthed along with the other files. This PR does however cause a potentially breaking change as the reactTypeDef has been removed as a public property. 
 
Tested locally and can confirm all files synth correctly and the project now builds with no errors.

Fixes: #1915

BREAKING CHANGE: `reactTypeDef` is no longer exposed on React project types, and `ReactTypeDef` and `ReactTypeDefOptions` are deprecated. This file is now created as a sample file so it will be managed by the user (and they can edit it, delete it, etc.) instead of by projen. Create a `SourceCode` file instead if you would like to use a managed react type definitions file in your project.

---
By submitting this pull request, I confirm that my contribution is made under the terms of the Apache 2.0 license.