feat(ObjectFile): `addToArray` escape hatch method (#1920)
The `addToArray` method allows values to be added to existing arrays within the object.

Unlike the `addOverride` method, `addToArray` works against `this.obj` rather than `this.rawOverrides`. It needs to do this in order to add to the array. Other suggestion approaches are welcome.

I don't really like the name `addToArray`, but can't think of anything better. Suggestions are welcome.

Any other feedback is also welcome.

---
By submitting this pull request, I confirm that my contribution is made under the terms of the Apache 2.0 license.