fix: restore ability to lint files in projenrc src directory (#1926)
Closes #1903 

The change in #1832 inadvertently removed required configuration for files on `projenrc` to be linted. This change restores that config by once again creating the `ProjenrcTs` component _after_  the `Eslint` component. It also adds `projenrc` as a `testDir`, so files will be linted by the eslint task.


This PR is aiming to resolve #1903 with minimal changes. I felt that the current implementation of the `Eslint` component is mixing responsibilities a bit (is it aware of projen, how much is it aware?) and makes some assumptions about what it thinks it knows about projen. **tl;dr I didn't really want to get into it.** But please let me know if you do want me to!


---
By submitting this pull request, I confirm that my contribution is made under the terms of the Apache 2.0 license.