fix(publisher): add manual publish condition
Adding a condition to the git publish command, which performs
the same check as the bump command. If the last commit was a release
commit, it doesn't make sense to try to publish again. Doing so
actually causes an error because bump is skipped which means there
is no version file to read.