fix(publisher): add manual publish condition (#1950)
Adding a condition to the git publish command, which performs
the same check as the bump command. If the last commit was a release
commit, it doesn't make sense to try to publish again. Doing so
actually causes an error because bump is skipped which means there
is no version file to read.

---
By submitting this pull request, I confirm that my contribution is made under the terms of the Apache 2.0 license.