feat: create sample test files only if jest is enabled (#1959)
Currently, TypeScript projects create a `test` sample directory with a `hello.test.ts` file in it. It doesn't make sense to create those files if `jest` is not enabled in the project. Furthermore, trying to add custom sample tests for other test runners such as `vitest` is impossible since the `SampleDir` is already instantiated.

---
By submitting this pull request, I confirm that my contribution is made under the terms of the Apache 2.0 license.