fix(circleci): add pipeline parameters object, remove parameters from workflow job (#1969)
Pipeline parameters were not defined according to the CircleCI spec.  This change fixes that.  See https://circleci.com/docs/2.0/reusing-config#using-the-parameters-declaration and https://circleci.com/docs/2.0/configuration-reference#jobs-in-workflow

fixes #1968 

BREAKING CHANGE: WorkflowJob no longer has a parameters key, and the parameters key in Job is now an object with `description?`, `type`, and `default?` keys.

---
By submitting this pull request, I confirm that my contribution is made under the terms of the Apache 2.0 license.