feat: add smithy build support to projen (#1970)
For folks using [smithy](https://awslabs.github.io/smithy/index.html) in their project, it would be great to support generation of `smithy-build.json` file using projen.

The types defined in this PR could be refined further for type safety but this PR is a starting point for those changes. [This](https://awslabs.github.io/smithy/1.0/guides/building-models/build-config.html#projections) was used as a reference for defining types in smithy-build.ts file.

Shoutout to @eduardomourar for collaborating with me on this!

---
By submitting this pull request, I confirm that my contribution is made under the terms of the Apache 2.0 license.