feat(awscdk): custom Lambda runtimes (#1982)
Allow to use custom Lambda runtimes for bundled functions.

Remove the "coupling" with the enum from `aws-cdk-lib.aws_lambda`
so the runtime can be updated to a newer runtime without having to
update the minimum version of the AWS CDK to depend on.

---
By submitting this pull request, I confirm that my contribution is made under the terms of the Apache 2.0 license.