feat(release): add option to set the minimal major version
This new option is really convenient when you have a generated project
that is self updating likee the cdktf pre-built providers. With the
current setup it's impossible to increase the major version without
manual intervention. Since breaking changes in a 0.x version only
result in a minor version increase there is no way to get to a
higher major version.