feat!: add Project.commitGenerated prop (#1972)
By default, files managed by projen are commited to the repository. That's not entirely necessary in some cases, and this commit allows customizing this behavior. By providing the Project.defaultCommitManagedFiles prop, we can tell projen whether to commit the files or not (if the File.commited prop is omitted).

The motivation for this change is to have cleaner repositories while maintaining the same developer experience.

---
By submitting this pull request, I confirm that my contribution is made under the terms of the Apache 2.0 license.