feat: optionally support comments in .json and in .jsonc files (#2045)
VS Code and other tools might accept comments in regular .json files.
In these situations, the preferred way of adding comments (like the Marker)
would be a comment instead of the `//` object property.

---
By submitting this pull request, I confirm that my contribution is made under the terms of the Apache 2.0 license.