feat(vscode): support VS Code workspace settings (#2047)
Settings keys languages are not further typed as there are A LOT of
them and plugins can add any new values to it.

Setting values are not typed as they are essentially of type JSON which is
difficult to express in TS and a jsii compatible way.

---
By submitting this pull request, I confirm that my contribution is made under the terms of the Apache 2.0 license.