feat: Adding notes about forcing release versions (#2070)
I struggled in the beginning understanding how I could "catch up" a project that was taking on projen to get version numbers to match up, so I added a section about it.

---
By submitting this pull request, I confirm that my contribution is made under the terms of the Apache 2.0 license.