fix(bump)!: respect existing tags for prerelease
This change fixes the behavior of the bump task when a prerelease
suffix is set, there are not yet any prerelease tags, but there are
other normal tags for the current major version. The current behavior
results in "first release" behavior. What should happen, and what this
change implements, is a correct semver increment into a new prerelease.

For example,

Given

- A major version of 1
- A previous tag of 1.2.0
- A new feat commit
- A prerelease setting of beta

Previously a bump would have resulted in 1.0.0-beta.0.
Now a bump will result in 1.3.0-beta.0, as expected.