fix(awscdk): Export `awscdk-deps-py` in index
The `AwsCdkDepsPy` class cannot be discovered by an external project as it's not exported in the module's `index.py`. I was able to confirm this in my external project by modifying the locally-installed `projen`: by updating `lib/awscdk/index.d.ts` and `lib/awscdk/index.js` to `export * from "./awscdk-deps-py", my project was able to find `AwsCdkDepsPy` and build. Removing those added exports made it unable to find it again.