feat(release): allow github oidc auth provider for aws code artifact
This closes #1313.

Empehmeral credentials are more secure than fixed ones. Prior to this
change, users had to have fixed secrets defined in Github for an AWS
access/secret key pair to publish to CodeArtifact. With this change, it
is now possible to use Github's OIDC provider to assume a role without
the need for any long-lived credentials.

Backwards compatibility is preserved and all existing users will be
unaffected unless they explicitly update the `authProvider` of
`codeArtifactOptions` to `GITHUB_OIDC`.

---
By submitting this pull request, I confirm that my contribution is made under the terms of the Apache 2.0 license.