fix: lambda function not bundled with tsconfig.dev.json (#2116)
If no `tsconfig.json` is present, esbuild defaults are used for bundling. 

This might be okay, but can yield inconsistent builds if `tsconfig.json` is not checked into the repository, like it is for `JsiiProject`s. E.g. `tsconfig.json` exists locally (due to an earlier run of jsii) and is used by local bundling. However in CI/CD it does not exist and bundling will create a different output (might be okay, assets are not checked in) and therefore a different hash (this is a problem with integration tests) causing the self-mutation step to file.


---
By submitting this pull request, I confirm that my contribution is made under the terms of the Apache 2.0 license.