fix(release): artifact permissions are not preserved in github actions (#2132)
Fixes #2103.

Due to a bug in GitHub actions, file mode permissions (such as `+x`) are lost between when the build artifact is uploaded at the end of the build job and download in each release job. This results in the published artifact tarballs having incorrect permissions. The main user impact is that bin scripts are no longer executable. NPM automatically fixes this during install, but other JSII toolchains, such as Python, do not.

See details of GitHub Actions bug [here](https://github.com/actions/upload-artifact/issues/38).

To workaround the issue, this PR backs up the permissions of the artifact tree to a file (`permissions-backup.acl`) prior to upload, then restores them after download using the Linux `getfacl` and `setfacl` CLI tools.


---
By submitting this pull request, I confirm that my contribution is made under the terms of the Apache 2.0 license.