fix: getfacl not found (#2136)
In #2132 we fixed an issue with file permissions not being preserved in GH Action workflows. The implementation uses the `acl` package to backup end restore permissions.  While we can ensure the package is installed in all default runners or containers, it might not be available in customized workflows.

This change marks the permissions backup and restore steps as allowed to fail. Since it is a workaround for GH Actions issue and only required in certain situations, we should not fail the whole build if the package is not available.

Fixes: #2134 

Additionally fixes an issue with invalid workflow files being generated when `continueOnError` is set on a Job Step.

---
By submitting this pull request, I confirm that my contribution is made under the terms of the Apache 2.0 license.