fix(projenrc): projenrc class does not create `--outdir` (#2138)
This PR fixes in #2099 non-working `--outdir` - option.
The problem was that `writeFileSync` creates a file only but requires that the folder which is passed by `--outdir` already exists.

Here we switched the lib from `fs` to `fs-extra` and changed to `outputFileSync` instead of `writeFileSync`.

fixes #2099 

---
By submitting this pull request, I confirm that my contribution is made under the terms of the Apache 2.0 license.