chore: re-write release docs (#2137)
[Rendered Version](https://github.com/mrgrain/projen/blob/docs/manual-release/docs/releases.md)

The current documentation for releases is unclear in parts and has received valuable, but somewhat disconnected additions addition over time. This re-write attempts to create a better narrative around releasing and versioning. Specifically I have added a section on the Initial development phase and moved some edge cases into a new FAQ section at the end of the page.

A major overhaul was done for local (aka manual) releases, which did not explain very well how they work. This especially confused users, either expecting `projen release` to publish their package or attempting to publish the repository directly. The updated docs clearly outline the current behavior and required manual steps to publish a package. It also explains why we set the version in `package.json` to `0.0.0`.

Please note that the additions for local releases are just to document the _current_ behavior better. It would be feasible and/or desirable to add new features that:
- add a publish task to run separately and/or as part of release
- keep the version in package.json in sync with tags

Also includes a small clarification for the `authorOrganization` flag and adds the correct default value for `releaseTagPrefix`.

Fixes: #1938

---
By submitting this pull request, I confirm that my contribution is made under the terms of the Apache 2.0 license.