chore: remove package overrides for `got` and `@types/responselike` (#2164)
Fixes: #2163

The underlying issue (https://github.com/sindresorhus/got/issues/2129) has been resolved and packages are again building correctly with the latest versions of the affected packages. This PR serves as proof of this.

---
By submitting this pull request, I confirm that my contribution is made under the terms of the Apache 2.0 license.