chore(docs): fix link to PomOptions API reference in docs for Java projects (#2212)
The current documentation for Java projects contains a dead link to the API reference for `PomOptions`.

This PR fixes the path to the linked Markdown file containing the API reference (`api/API.md`).

---
By submitting this pull request, I confirm that my contribution is made under the terms of the Apache 2.0 license.