fix(javascript): Remove `--all` argument to jest CLI
This argument does not exist and hasn't existed since at least 22.x:

https://jestjs.io/docs/cli

https://archive.jestjs.io/docs/en/22.x/cli

This argument was added in Dec 2020 by a larger commit adding build tasks:
95112272c2b192144293c1064c77b2d8da354b8f

And before that, the whole file did not exist and the project has no
mention of `--all`. So I cannot deduce the original reason for adding
this argument.