feat(tasks): allow passing arguments to tasks (#2246)
Closes: #356

Basic implementation of allowing arguments to be passed through to tasks. Currently not used by any project tasks, but this can be added in future.

Could be extended in future to support well defined (positional) args and options.

---
By submitting this pull request, I confirm that my contribution is made under the terms of the Apache 2.0 license.