chore(dependencies): Update peter-evans/create-pull-request to v4 (#2327)
The `peter-evans/create-pull-request` action has a v4 available, which
is the version actively receiving updates. v4 has been updated to work
with the new version of NodeJS as well as to use the new version of
`@actions/core`, which uses the new `set-output` replacement. This
update will reduce the number of warnings emitted when the
`upgrade-main` workflow runs.

Besides the NodeJS version change (which the versions of
`actions/checkout`, etc already have made), the only meaningful breaking
change in v4 was the way `add-paths` argument works, which is not used
in these workflows.

- [v4.0.0 Release Notes](https://github.com/peter-evans/create-pull-request/releases/tag/v4.0.0)
- [v4.1.4 Release Notes](https://github.com/peter-evans/create-pull-request/releases/tag/v4.1.4)
- [Sample workflow run](https://github.com/projen/projen/actions/runs/3784184375)

---
By submitting this pull request, I confirm that my contribution is made under the terms of the Apache 2.0 license.