fix(jsii): run workflow setup steps for language jobs (#2311)
Fix #1799 #1006

Allow workflow setup steps to be used by extensions of node-project. 

This change will enable for example a step that sets NPM authentication for private repos before build and publish.

---
By submitting this pull request, I confirm that my contribution is made under the terms of the Apache 2.0 license.