chore(javascript): reduce the scope of the upgrade dependencies token (#2312)
The second step of this workflow that handles opening the pull request
is unnecessarily broad. Because the `PROJEN_GITHUB_TOKEN` (or the
otherwise configured token) will have `contents:write` and
`pull-requests:write` and that token is the one that is used to actually
push the content and open the pull request (as part of the
`peter-evens/create-pull-request` action), the `GITHUB_TOKEN` can be
used exclusively to checkout the branch with `contents:read`.

This change reduces the permissions of the `GITHUB_TOKEN` to the minimal
amount required and also reduces the places the configured token is used
to creating the pull request.

---
By submitting this pull request, I confirm that my contribution is made under the terms of the Apache 2.0 license.