chore(projects): Set environment variable in VM when creating new project (#2351)
Currently it's not really possible for external project types to distinguish between when a project is being created versus routine synth's. 

Being aware of this can be useful for certain types of project customizations, especially given external projects' limited ability to customize the "projen new ..." experience. 

This PR sets an environment variable (`PROJEN_CREATE_PROJECT=true`) in the VM used to create the project, thus allowing external project types to be aware of the nature of the current synthesis.

---
By submitting this pull request, I confirm that my contribution is made under the terms of the Apache 2.0 license.