feat(github): allow specifying permission scope for app token (#2380)
This updates `tibdex/github-app-token` to v1.7.0; newer versions of this
action support specifying the specific permissions that should be
granted to the created token.

This creates a new interface for `AppPermissions` following the
documentation to [create an installation access token][1]. Additionally,
these permissions can be `read`, `write`, or `admin`; however, there is
not a `none` options, so the `JobPermission` enum cannot be reused and a
new `AppPermission` enum is defined.

A test is also added to test specifying this.

Ideally the _default_ for `permissions` would be to just set
`pullRequests` and `contents` to `WRITE`; however, if users have added
steps onto the workflow that require additional permissions, those may
break if the default permissions are changed from "all available" to
"just PRs and contents".

[1]: https://docs.github.com/en/rest/apps/apps?apiVersion=2022-11-28#create-an-installation-access-token-for-an-app

---
By submitting this pull request, I confirm that my contribution is made under the terms of the Apache 2.0 license.