chore(docker-compose): split docker-compose and move to own submodule (#2403)
As proposed by me in #2397, I'd like to implement networks into the `DockerCompose` component, but as the file was already large I wanted to take the opportunity to refactor it beforehand.

Therefore, I:
1. Kept all the logic and code the same, but moved the docker-compose component to an own directory
2. Split the larger file into multiple smaller ones, by resource type and purpose, to make it easier to navigate

I don't expect this to be a breaking change. All interfaces are still exported to the same path and with the same name. I outlined that in the above mentioned issue.

---
By submitting this pull request, I confirm that my contribution is made under the terms of the Apache 2.0 license.