feat: convert TypescriptConfig to Component and allow comments (#2409)
This PR converts `TypescriptConfig` class into a subclass of `Component` and proposes omitting default values instead of hardcoding them.

The motivation for inheriting from `Component` is the same as per #2395 but as we can have several instances of `TypescriptConfig` per project we are not providing a static `.of` method. Yet this change allows us to easily feature-detect the presence of `TypescriptConfig` inside a project. This is incredibly handy in the case of using this class in some non-standard projects (i.e. Typescript monorepo root or as a base for `jsconfig.json`).

This PR also fixes the fact that `tsconfig.json` support comments. [1]

We also added a test for this component (we didn't have any) where we are loading generated config using the API of Typescript itself - that will catch any possible mistakes in `compilerOptions` as TS throws on unknown values.

[1] https://github.com/microsoft/TypeScript/pull/5450

---
By submitting this pull request, I confirm that my contribution is made under the terms of the Apache 2.0 license.