feat(typescript): support `noImplicitOverride` (#2442)
Allows to require use of the `override` keyword to indicate that a member overrides a declaration from a parent type, improving code readability and providing additional protection against typos and rename churn.

---
By submitting this pull request, I confirm that my contribution is made under the terms of the Apache 2.0 license.