fix(new): cannot create a new project if the path has a space (#2444)
Strings that might contain spaces need to be quoted when passed into `exec`.
This fixes the issue when the base path that `projen new` is run from contains a space.

Fixes #2443

---
By submitting this pull request, I confirm that my contribution is made under the terms of the Apache 2.0 license.