feat(github): allow workflow action versions to be set (#2439)
It is a good security practice to lock versions of GitHub Actions down to explicit commit hashes.
Ideally projen would manage and upgrade GitHub Actions versions like any other dependency.
However in absence of this we need a mechanism for users to change the version of workflows that come with projen.

This change implements to smallest possible solution.
Users can define explicit overrides for each used action if desired.
The API allows to override a specific action version, which can be useful if an action has breaking changes between major versions.
Alternatively all occurrences of an action can be overridden, irregardless of the action version.

Workflow creators continue to define actions as regular strings.
The recommendation would be to use commit hashes if they will be kept up to date, and major versions otherwise.
Sensitive workflow actions should always use commit hashes.

Closes #1844 and closes #1702

---
By submitting this pull request, I confirm that my contribution is made under the terms of the Apache 2.0 license.