feat: add support for Git LFS (#2455)
Git LFS needs all workflow files to be updated, so it needs to  go into projen core.

API looks like this:

```ts
new Project({
  githubOptions {
    lfsPatterns: ['*.bin'],
  }
});
```

This does 2 things:

* Add the right configuration lines to `.gitattributes`
* Adds `lfs: true` the options of all `checkout@v3` GitHub Actions. 

Strictly speaking, the modeling is not entirely correct: LFS patterns should be an attribute of `gitOptions`, not `githubOptions`, but there doesn't seem to be such a component, so this seems like the next best place to put it.

---

By submitting this pull request, I confirm that my contribution is made under the terms of the Apache 2.0 license.