feat(node-package): Added ability to set pnpm version for action-setup (#2482)
There is currently an [issue](https://github.com/pnpm/pnpm/issues/5985) with PNPM in which ca:login fails to find the correct npm registry value for authentication in the latest version.

To mitigate this we want to be able to specify the version of PNPM to use in the Github workflow. 
This may help with future incompatibilities as well.

~This change also updates the configuration of **pnpm/action-setup** to use `latest` (currently 7.29.0) instead of being hard coded to 7~

Upon code review feedback this update now sets the default to version 7.   

---
By submitting this pull request, I confirm that my contribution is made under the terms of the Apache 2.0 license.