feat(github): support all crud operations for github-workflow jobs (#2486)
In my custom projects I wanted to generate standard workflows that are also dynamic based on which components is enabled. While still allowing users to overwrite these defaults in case it is not to their liking. I saw this is not possible because the job property is private and we only have add functions in the workflow class. 

I've added the getJob, updateJob(s), removeJob functions so users can do this.

---
By submitting this pull request, I confirm that my contribution is made under the terms of the Apache 2.0 license.